Revert "February 13th, 2025 Update"
This reverts commit 9098990d9fe59c6496e3a6f63f986a9eee425014.

+++ selfdrive/frogpilot/ui/qt/offroad/sounds_settings.cc
@@ -57,7 +57,7 @@ FrogPilotSoundsPanel::FrogPilotSoundsPanel(FrogPilotSettingsWindow *parent) : Fr
       });
       soundsToggle = alertVolumeControlToggle;
     } else if (alertVolumeControlKeys.find(param) != alertVolumeControlKeys.end()) {
-      std::map<float, QString> volumeLabels;
+      std::map<int, QString> volumeLabels;
       for (int i = 0; i <= 101; ++i) {
         volumeLabels[i] = i == 101 ? tr("Auto") : i == 0 ? tr("Muted") : QString::number(i) + "%";
       }
